Allow float heart rate values; add parsing tests

+++ app/src/main/java/com/example/smartwatchhapticsystem/controller/BluetoothConnectionManager.java
@@ -287,7 +287,7 @@ public class BluetoothConnectionManager {
         // Validate and deliver heart rate value
         if (dataMap.containsKey("Value")) {
             try {
-                Integer.parseInt(dataMap.get("Value")); // Validate numeric format
+                Double.parseDouble(dataMap.get("Value")); // Validate numeric format
 
                 if (listener != null) {
                     handler.post(() -> listener.onReceived(dataMap));
